requirements.txt
numpy
pandas
scikit-learn
tensorflow
google-generativeai
flask